# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Edge Platform industrial IoT / HMI cloud API.",
        "title": "Edge Platform API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/system/config": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "新增配置",
                "parameters": [
                    {
                        "description": "配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sysconfig.configRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysconfig.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/config/batch-delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "批量删除配置",
                "parameters": [
                    {
                        "description": "配置 ID 列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sysconfig.idsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysconfig.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/config/key/{key}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "按配置键查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置键",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysconfig.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/config/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "配置分页",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysconfig.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/config/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "配置详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysconfig.ApiEnvelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "修改配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sysconfig.configRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysconfig.ApiEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "删除配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysconfig.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/config/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "修改配置状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sysconfig.statusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysconfig.ApiEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "sysconfig.ApiEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "sysconfig.configRequest": {
            "type": "object",
            "properties": {
                "configKey": {
                    "type": "string"
                },
                "configName": {
                    "type": "string"
                },
                "configType": {
                    "type": "string"
                },
                "configValue": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "valueType": {
                    "type": "string"
                }
            }
        },
        "sysconfig.idsRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "sysconfig.statusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}